ok turns out it wasn't really working, it only works for way too large contact distance thresholds which wont fly.. and now I'm finding out that if I move from maxiter to however many interations it takes, it's never finding a solution unlike the python version. Rn I'm working on switching over to doubles to see if that fixes anything

+++ src/utils/sdf.h
@@ -4,10 +4,10 @@
 #include <memory>
 
 struct SDF {
-	virtual float distance(Eigen::Vector3f& point) = 0;
+	virtual double distance(Eigen::Vector3f& point) = 0;
 	virtual Eigen::Vector3f distanceGrad(Eigen::Vector3f& point) = 0;
 	virtual Eigen::Matrix3f distanceHess(Eigen::Vector3f& point) = 0; 
-	virtual float ccd(Eigen::Vector3f& point, Eigen::Vector3f& partialSearchDir) = 0;
+	virtual double ccd(Eigen::Vector3f& point, Eigen::Vector3f& partialSearchDir) = 0;
 };
 
 struct PlaneSDF : SDF {
@@ -15,16 +15,16 @@ struct PlaneSDF : SDF {
 	Eigen::Vector3f planePoint;
 	
 	PlaneSDF(Eigen::Vector3f normal, Eigen::Vector3f point);
-	float distance(Eigen::Vector3f& point) override;
+	double distance(Eigen::Vector3f& point) override;
 	Eigen::Vector3f distanceGrad(Eigen::Vector3f& point) override;
 	Eigen::Matrix3f distanceHess(Eigen::Vector3f& point) override; 
-	float ccd(Eigen::Vector3f& point, Eigen::Vector3f& partialSearchDir) override;
+	double ccd(Eigen::Vector3f& point, Eigen::Vector3f& partialSearchDir) override;
 };
 
 // struct MeshSDF : SDF {
 // 	std::shared_ptr<Mesh> mesh;
 
-// 	float distance(Eigen::Vector3f point) override;
+// 	double distance(Eigen::Vector3f point) override;
 // 	Eigen::Vector3f distanceGrad(Eigen::Vector3f point) override;
 // 	Eigen::Matrix3Xf distanceHess(Eigen::Vector3f point) override; 
 // };
